Apply a simple  middlware logger
Redux provides a middlware concept right between the action is dispacthed and reaching the reducer.
Middlware is a term used for functions or code in general you hook in a process which then gets executed as a part of the process without stopping it.
We can use the redux middleware to do something with the action before the action reaches the reducer.

Here we want to add a logger to log the action and the next state.