Create action creators
Ultimately we want to use asychronous code. We cannot use asychronous code in reducers. The reducer function has to run sychronously. The reducer is going to be already  executed by the time the asychronous code takes place.So we need to create action creators where we can have the asychronous code in there.