Add redux-thunk package which is a middleware to help handle asychronous code
It is a library that adds a middleware to the project which allows your action creators to not return the action itself but return a function which will eventually dispatch an action. With this we can run an asychronous code because the eventually dispatch part is the part that may run asychronously.