Restruct Actions
Split up action creators over multiple files.
For this demo project is overkill however it is needed for bigger projects because it allows you to have files with a few action creators in each file which it makes it easier to find that action creator you want to change or check.